move chatUtils.ts to /resources
