+++ src/components/resources/chatUtils.ts
